server/requirements.txt
alembic==1.13.3
amqp==5.3.1
aniso8601==9.0.1
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
cachelib==0.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.55.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.5
MarkupSafe==2.1.5
pillow==10.4.0
prompt_toolkit==3.0.48
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
razorpay==1.4.2
redis==5.2.0
requests==2.32.3
setuptools==75.5.0
six==1.16.0
SQLAlchemy==2.0.35
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.0.4
zopfli==0.2.3.post1
